+++ src/meeting_recorder/__init__.py
@@ -1,3 +1,3 @@
 """Meeting Recorder App."""
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"